moving classes to libs folder
